Start game when handshake happens to fix the bug
The server will send the current state to the clients,
when the server receives the 'game-start' message by client's clicking play button.
When the server sends 'game-start' to client, it will creates puzzle-board.
So, the puzzle will not be changed by the other client's join.
(Before, it was changed by the new client's new puzzle game-state)