Fix bug: Create puzzle-completion-message once when game is created
So that the puzzle-completion-message is not displayed down far,
when players visit the ranking dashboard and come back to game.
Still it makes the black color div below the game puzzle page,
But at least, puzzle-completion-message(including ranking button, play button)are
not created the newly created (black colored)div below the puzzle page.